ChatDoc/requirements.txt
streamlit 
embedchain
watchdog
langchain_community 
tiktoken 
angchain-openai 
langchainhub 
chromadb 
langchain
youtube-transcript-api 
pytube
bs4
PyPDF2
faiss-cpu
pdfminer